requirements.txt
ipykernel
pandas
psycopg2-binary
sqlalchemy
